requirements.txt
streamlit==1.45.0
openai==1.68.2
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn==0.34.0
httpx==0.28.1
